requirements.txt
# core
numpy>=1.24.0
scipy>=1.10.0
pyyaml>=6.0

# visualization & dashboard
matplotlib>=3.7.0
streamlit>=1.28.0
plotly>=5.15.0

# ml - will use later
scikit-learn>=1.3.0

# stm32 serial communication
pyserial>=3.5

# parallel processing
# using built-in multiprocessing + concurrent.futures
